pyproject.toml
[project]
name = "pupil-labs-mar-common"
description = "Marc's repo of commonly reused things!"
authors = [{ name = "Pupil Labs GmbH", email = "info@pupil-labs.com" }]
license = { text = "Proprietary" }
readme = "README.md"
requires-python = ">=3.10, <=3.11"
keywords = []
dynamic = ["version"]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "numpy",
    "pupil-labs-neon-usb>=0.0.1.post2",
    "pupil-labs-realtime-api>=1.7.3",
    "pyside6>=6.10.1",
    "python-dotenv>=1.2.1",
    "qimage2ndarray>=1.10.0",
]

[project.optional-dependencies]
examples = [
    # "opencv-python",
    # "pandas",
    # "tqdm",
]

[project.urls]
Homepage = "https://pupil-labs.github.io/mar-common"
Documentation = "https://pupil-labs.github.io/mar-common"
Repository = "https://github.com/pupil-labs/mar-common"
Issues = "https://github.com/pupil-labs/mar-common/issues"


[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pre-commit>=2.20.0",
    "mypy>=0.991",
    "deptry>=0.20.0",
    "tox-uv>=1.11.3",
    "pytest-cov>=4.0.0",

    # stubs
    "opencv-stubs",

    # docs
    "markdown-callouts>=0.4",
    "markdown-exec>=1.8",
    "mike",
    "mkdocs>=1.6",
    "mkdocs-coverage>=1.0",
    "mkdocs-gen-files>=0.5",
    "mkdocs-git-revision-date-localized-plugin>=1.2",
    "mkdocs-literate-nav>=0.6",
    "mkdocs-material>=9.5",
    "mkdocs-minify-plugin>=0.8",
    "mkdocstrings[python]>=0.25",
    # YORE: EOL 3.10: Remove line.
    "tomli>=2.0; python_version < '3.11'",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
log_format = '%(asctime)s %(levelname)7s %(name)-60s | %(message)-100s "%(pathname)s:%(lineno)s"'

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["pupil_labs"]

[tool.deptry.per_rule_ignores]
DEP001 = ["pupil_labs", "cv2"]
DEP002 = ["opencv-python"]
DEP003 = ["pupil_labs", "typing_extensions"]

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build.targets.wheel]
packages = ["src/pupil_labs"]
